requirements.txt
fastapi==0.115.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.9.0
boto3==1.35.36
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
stripe==8.0.0
python-multipart==0.0.9
openai==1.51.2
google-genai==0.3.0
google-auth
google-auth-oauthlib
google-api-python-client
